Move Failed Import Detection into a feature
