+++ src/ServiceControl/ServiceControl.csproj
@@ -440,7 +440,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Infrastructure\RavenDB\RavenBootstrapper.cs" />
     <Compile Include="MessageAuditing\ProcessingStatisticsEnricher.cs" />
-    <Compile Include="Operations\AttemptToReImportFailedMessages.cs" />
+    <Compile Include="Operations\DetectFailedMessageImportsFeature.cs" />
     <Compile Include="Recoverability\Grouping\Archiving\ArchiveAllInGroup.cs" />
     <Compile Include="Recoverability\Grouping\Archiving\ArchiveAllInGroupHandler.cs" />
     <Compile Include="Recoverability\Grouping\ClassifyFailedMessageEnricher.cs" />
